CI: Move disk tests into separate file
